Create README.md
This is a utility application currently under development. Its goal is to unite several existing scripts I had into a simple and easy to use tool. All the scripts are getting facelifts as I make this so there will be bugs. This is free to use but use at your own discretion. 

Currently added:
KeyLogger - Saves to a file called log.txt. Use the PageDown key to stop it from logging anymore. This key can be changed in the resource file.

In the works:
Password manager - Upload and encrypt user creds to a db of your choice

Planned:
Note taker - save quick notes and back them up to an online storage.